MultiLabelSegmentation/requirements.txt
matplotlib
numpy
scikit_learn
pymaxflow
